mount partition for sieve scripts
James already setsup a SieveMailet that is filtering mails. If filtering fails (i.e. if there are
no scripts), the mail is added to INBOX. James expects the sieve scripts in a specific location.
The filesystem-partition mounted puts this location into the content tree, such that sieve scripts
can be easily edited online. Unfortunately, it's not (yet) possible to create a git partition at
an arbitary location on the server; could be changed later to have sieve scripts under version
control