fixes delivering mails to multiple local recipients with active sieve script
for each recipient the mail is handed to sieve's fileinto action. this action
expects exactly one recipient. In case of many recipients the SieveMailboxMailet
does not reset the recpient lists -- although it is executing storeMail() with the
same mail for each recipient. This patch just overrides the sieveMessage() method
to set the recipient list to the given recpient.